refactor: 重构修改章节控件，改为组件化
